- Prevented move to processed - Added check for existence of database
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/LOKI@2173 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ branches/1.0/updater/bio_loader.py
@@ -164,8 +164,9 @@ def LoadKB(dbFilename, kbLoads, doReset = False, refreshEnsembl = False):
 	except Exception, e:
 		#unhandled exception means something went VERY wrong
 		# if it was a new db, just delete it!
-		if not db_exists:
-			os.chdir(cwd)
+		
+		os.chdir(cwd)
+		if not db_exists and os.path.exists(dbFilename):
 			os.remove(dbFilename)
 
 		raise
